Fix some bugs in BFS2::findBestRoots()

+++ Morph/BFS2.h
@@ -37,7 +37,7 @@ public:
 	RoadGraph* copyRoads(RoadGraph* roads, BFSTree* tree, int num);
 
 	void findBestRoots(RoadGraph* roads1, BFSTree* tree1, RoadGraph* roads2, BFSTree* tree2, RoadVertexDesc& root1, RoadVertexDesc& root2);
-	float computeUnbalanceness(RoadGraph* roads1,  BFSTree* tree1, RoadVertexDesc node1, RoadGraph* roads2,  BFSTree* tree2, RoadVertexDesc node2);
+	int computeUnbalanceness(RoadGraph* roads1,  BFSTree* tree1, RoadVertexDesc node1, RoadGraph* roads2,  BFSTree* tree2, RoadVertexDesc node2);
 
 	void createRoads1();
 	void createRoads2();
